Added data repositories and enums

+++ src/main/java/Model/StockItem.java
@@ -6,13 +6,14 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
 import java.util.List;
+import java.math.BigDecimal;
 @Entity
 @Table(name="StockItem")
 public class StockItem extends Product{
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long stockID;
-    private double stockPrice;
+    private BigDecimal stockPrice;
     private String stockColour;
     private String stockSize;
     private int stockQty;
@@ -22,7 +23,7 @@ public class StockItem extends Product{
     }
 
     // Stock Item Constructor
-    public StockItem(long stockID, int productID, String productName, String productBrand, String productDecade, String productDescription, double importPrice, double productPrice, double discountPrice, List<String> productColour, List<String> productSize, int productQty, String supplierCode, double stockPrice, String stockColour, String stockSize, int stockQty){
+    public StockItem(long stockID, int productID, String productName, String productBrand, String productDecade, String productDescription, BigDecimal importPrice, BigDecimal productPrice, BigDecimal discountPrice, List<String> productColour, List<String> productSize, int productQty, String supplierCode, double stockPrice, String stockColour, String stockSize, int stockQty){
         super(productID, null, productName, productBrand, productDecade, productDescription, importPrice, productPrice, discountPrice, productColour, productSize, productQty, supplierCode);
         this.stockID = stockID;
         this.stockPrice = stockPrice;
@@ -34,8 +35,8 @@ public class StockItem extends Product{
     //Getters and setters
     public long getStockID(){return stockID;}
     public void setStockID(long stockID) {this.stockID = stockID;}
-    public double getStockPrice(){return stockPrice;}
-    public void setStockPrice(double stockPrice) {this.stockPrice = stockPrice;}
+    public BigDecimal getStockPrice(){return stockPrice;}
+    public void setStockPrice(BigDecimal stockPrice) {this.stockPrice = stockPrice;}
     public String getStockColour(){return stockColour;}
     public void setStockColour(String stockColour) {this.stockColour = stockColour;}
     public String getStockSize(){return stockSize;}
